[#4][REFACTOR] Refactored class & class name
